pip_requirements.txt
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.22.0
backports.shutil-get-terminal-size==1.0.0
cffi==1.10.0
cryptography==1.9
cycler==0.10.0
dcmstack==0.7.0.dev0
decorator==4.0.11
dicom==0.9.9.post1
enum34==1.1.6
functools32==3.2.3.post2
idna==2.5
ipaddress==1.0.18
ipython==5.4.0
ipython-genutils==0.2.0
line-profiler==2.0
matplotlib==2.0.2
mpmath==0.19
nibabel==2.1.0
nilearn==0.3.0
nipy==0.4.1
nose==1.3.7
numpy==1.12.1
packaging==16.8
paramiko==2.1.2
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
profilehooks==1.9.0
progressbar2==3.30.2
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyasn1==0.2.3
pycparser==2.17
pydicom==1.0.0a1
pyftpdlib==1.5.2
pygame==1.9.3
Pygments==2.2.0
pylibssh2==1.0.1
pyparsing==2.2.0
pyserial==3.3
python-dateutil==2.6.0
python-utils==2.1.0
pytz==2017.2
scandir==1.5
scikit-learn==0.18.1
scipy==0.19.0
sftpserver==0.3
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
subprocess32==3.2.7
sympy==1.0
traitlets==4.3.2
wcwidth==0.1.7
